+++ hooks/post_gen_project.py
@@ -3,9 +3,11 @@ Does the following:
 
 1. Inits git if used
 2. Deletes dockerfiles if not going to be used
+3. Deletes config utils if not needed
 """
 from __future__ import print_function
 import os
+import shutil
 from subprocess import Popen
 
 # Get the root project directory
@@ -43,11 +45,23 @@ def remove_docker_files():
             PROJECT_DIRECTORY, filename
         ))
 
+def remove_viper_files():
+    """
+    Removes files needed for viper config utils
+    """
+    shutil.rmtree(os.path.join(
+        PROJECT_DIRECTORY, "config"
+    ))
+
 # 1. Remove Dockerfiles if docker is not going to be used
 if '{{ cookiecutter.use_docker }}'.lower() != 'y':
     remove_docker_files()
 
-# 2. Initialize Git (should be run after all file have been modified or deleted)
+# 2. Remove Dockerfiles if docker is not going to be used
+if '{{ cookiecutter.use_viper_config }}'.lower() != 'y':
+    remove_viper_files()
+
+# 3. Initialize Git (should be run after all file have been modified or deleted)
 if '{{ cookiecutter.use_git }}'.lower() == 'y':
     init_git()
 else: