+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All enhancements and patches to Cookiecutter Golang will be documented in this f
 - git use now optional as well
 - added authors to credit contributors as well as general guidelines for contributing
 - added testing
+- viper config is now optional
 
 ## [2016-12-20]
 ### changed